Jetpack: Break out threat history list and use on scan page on wpcom
