+++ client/landing/jetpack-cloud/sections/scan/index.js
@@ -42,6 +42,17 @@ export default function () {
 			makeLayout,
 			clientRender
 		);
+
+		page(
+			'/scan/:site/:filter?',
+			siteSelection,
+			navigation,
+			scan,
+			wrapInSiteOffsetProvider,
+			showUpsellIfNoScan,
+			makeLayout,
+			clientRender
+		);
 	} else {
 		page( '/scan*', () => page.redirect( '/' ) );
 	}
