+++ client/landing/jetpack-cloud/sections/scan/main.tsx
@@ -6,6 +6,7 @@ import { connect } from 'react-redux';
 import { Button, ProgressBar } from '@automattic/components';
 import { translate } from 'i18n-calypso';
 import { flowRight as compose } from 'lodash';
+import config, { isEnabled } from 'config';
 
 /**
  * Internal dependencies
@@ -16,7 +17,6 @@ import SecurityIcon from 'components/jetpack/security-icon';
 import ScanPlaceholder from 'components/jetpack/scan-placeholder';
 import ScanThreats from 'components/jetpack/scan-threats';
 import { Scan, Site } from 'landing/jetpack-cloud/sections/scan/types';
-import { isEnabled } from 'config';
 import Gridicon from 'components/gridicon';
 import Main from 'components/main';
 import SidebarNavigation from 'my-sites/sidebar-navigation';
@@ -31,6 +31,7 @@ import contactSupportUrl from 'lib/jetpack/contact-support-url';
 import { recordTracksEvent } from 'state/analytics/actions';
 import { triggerScanRun } from 'lib/jetpack/trigger-scan-run';
 import { withApplySiteOffset, applySiteOffsetType } from 'components/jetpack/site-offset';
+import ThreatHistoryList from 'components/jetpack/threat-history-list';
 
 /**
  * Style dependencies
@@ -229,6 +230,14 @@ class ScanPage extends Component< Props > {
 		return this.renderScanOkay();
 	}
 
+	maybeRenderHistory() {
+		const { filter } = this.props;
+
+		const isJetpackCom = !! config( 'env_id' ).includes( 'jetpack-cloud' );
+
+		return ! isJetpackCom && <ThreatHistoryList { ...{ filter } } />;
+	}
+
 	render() {
 		const { siteId } = this.props;
 		if ( ! siteId ) {
@@ -241,6 +250,7 @@ class ScanPage extends Component< Props > {
 				<QueryJetpackScan siteId={ siteId } />
 				<PageViewTracker path="/scan/:site" title="Scanner" />
 				<div className="scan__content">{ this.renderScanState() }</div>
+				{ this.maybeRenderHistory() }
 			</Main>
 		);
 	}
